Add simple alerts

+++ src/components/CartPage.js
@@ -6,8 +6,11 @@ function CartPage() {
     const dispatch = useDispatch();
     const cartItems = useSelector(state => state.cart);
 
-    function removeCartItem(index) {
-        dispatch(removeItem({ index }));
+    function removeCartItem({ name, index }) {
+        if (window.confirm(`Are you sure you want to delete item: ${name}?`)) {
+            dispatch(removeItem({ index }));
+            alert(`Item by the name of: ${name} was deleted, come back anytime!`);
+        }
     }
 
     return cartItems.length ? (
@@ -18,9 +21,12 @@ function CartPage() {
                 {cartItems.map((cartItem, index) => (
                     <li key={`${cartItem.name}-${index}`} className="cart-item">
                         <h3 className="cart-item__title">
-                            {cartItem.name.toUpperCase()} ({cartItem.quantity})
+                            {cartItem.name} ({cartItem.quantity})
                         </h3>
-                        <button className="cart-item__button" onClick={() => removeCartItem(index)}>
+                        <button
+                            className="cart-item__button"
+                            onClick={() => removeCartItem({ name: cartItem.name, index })}
+                        >
                             Delete item
                         </button>
                     </li>
